NM27: 16-bit multiplication overflow in deleverage fee guard can brick emergency deleverage

+++ src/public/administration/write/OnlyEmergencyDeleverager.sol
@@ -42,7 +42,8 @@ abstract contract OnlyEmergencyDeleverager is
         nonReentrant
     {
         require(
-            deleverageFeeDividend * maxDeleverageFeeDivider <= deleverageFeeDivider * maxDeleverageFeeDividend,
+            uint256(deleverageFeeDividend) * maxDeleverageFeeDivider
+                <= uint256(deleverageFeeDivider) * maxDeleverageFeeDividend,
             ExceedsMaxDeleverageFee(
                 deleverageFeeDividend, deleverageFeeDivider, maxDeleverageFeeDividend, maxDeleverageFeeDivider
             )
